api corrected/details view added

+++ brewers-supporter/src/main/java/pl/brewers/supporter/brewerssupporter/dto/HoopingDataRequestDTO.java
@@ -4,12 +4,13 @@ import lombok.Builder;
 import lombok.Data;
 import pl.brewers.supporter.brewerssupporter.model.HoopingIngredient;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 @Builder
 public class HoopingDataRequestDTO {
-    private double gravity;
-    private double amount;
+    private BigDecimal gravity;
+    private BigDecimal amount;
     private List<HoopingIngredient> hoopingIngredients;
 }
