+++ brewers-supporter/src/main/java/pl/brewers/supporter/brewerssupporter/dto/HoopingDataResponseDTO.java
@@ -3,8 +3,10 @@ package pl.brewers.supporter.brewerssupporter.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class HoopingDataResponseDTO {
-    private double ibu;
+    private BigDecimal ibu;
 }
