+++ brewers-supporter/src/main/java/pl/brewers/supporter/brewerssupporter/dto/MashDataResponseDTO.java
@@ -3,8 +3,10 @@ package pl.brewers.supporter.brewerssupporter.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class MashDataResponseDTO {
-    private double gravity;
+    private BigDecimal gravity;
 }
